==> Project Code/1. Tweet Collection.ipynb <==
import snscrape.modules.twitter as sntwitter
import pandas as pd
from datetime import datetime, timedelta
import time

# Set search parameters
keyword = 'ukraine'
hashtag = ['#ukraine']#, '#ukrainewar', '#UkraineRussiaWar️']#, '#putinwarcriminal', '#zelenskywarcriminal', '#standwithrussia', '#istandwithrussia', '#StandWithUkraine', '#IStandWithUkraine']
start_date = datetime(2022, 1, 1)
end_date = datetime(2022, 12, 31)
delta = timedelta(days=1)

lang_code = "en"
follower_threshold = 100
tweet_limit = 100

t0 = time.time()
tweets = []

while start_date <= end_date:
    for tag in hashtag:
        query = f'{tag} lang:{lang_code} since:{start_date.date()} until:{(start_date + delta).date()} exclude:retweets exclude:replies'
        for i, tweet in enumerate(sntwitter.TwitterSearchScraper(query).get_items()):
            if tweet.user.followersCount >= follower_threshold:
                tweets.append({
                    'id': tweet.id,
                    'date': tweet.date,
                    'user': tweet.user.username,
                    'user_followers': tweet.user.followersCount,
                    'user_created': tweet.user.created,
                    'rendered_content': tweet.renderedContent,
                    'likes': tweet.likeCount,
                    'retweets': tweet.retweetCount,
                    'replies': tweet.replyCount,
                    'quoteCount': tweet.quoteCount,
                    'hashtags': tweet.hashtags,
                    'lang': tweet.lang,
                    'media': tweet.media,
                    'mentionedUsers': tweet.mentionedUsers,
                })
            if i >= tweet_limit:
                break
    start_date += delta
    
t1 = time.time()
print(round(t1 - t0), "seconds")

df = pd.DataFrame(tweets)
len(df)

df.head()

df.to_csv("Ukraine Tweets Unlimited.csv", index = False)
