==> Project Code/2. Feature Extraction and Processing.ipynb <==
## Preprocessing and Feature Extraction

import pandas as pd

df = pd.read_csv("../Data/Ukraine Tweets.csv")

## Initial Preprocessing

Check for duplicates

df = df.drop_duplicates(subset = 'id')

Precounting of features: Length, Hashtags, URLs and Mentions

df['tweet_length'] = df['rendered_content'].apply(len)

df['num_mentions'] = df['rendered_content'].apply(lambda x: x.count('@'))

df['num_hashtags'] = df['rendered_content'].apply(lambda x: x.count('#'))

df['num_urls'] = df['rendered_content'].apply(lambda x: x.count('https'))

Steps taken:
- Converting emojis to text
- We decide to remove all the mentions and hashtagged words, as these will be analysed separately
- Remove Links, as these don't contribute to SA
- Conducting the SA on our preprocessed data

Creating a new column so that we can see the adjusted tweet and original versiom

df.insert(loc=6,
          column='Adjusted Tweet',
          value=df['rendered_content'])

Converting emojis to text

import emoji

def demote(text):
    text = emoji.demojize(text)
    return text

df['Adjusted Tweet'] = df['Adjusted Tweet'].apply(demote)

Removing mentions and hashtagged words

import re

def remove_mentions_hashtags(text):
    text = re.sub("@[A-Za-z0-9_]+","", text)
    text = re.sub("#[A-Za-z0-9_]+","", text)
    return text

df['Adjusted Tweet'] = df['Adjusted Tweet'].apply(remove_mentions_hashtags)

Removing links

def remove_links(text):
    text = re.sub('http://\S+|https://\S+', '', text)
    text = re.sub('http[s]?://\S+', '', text)
    text = re.sub(r"http\S+", "", text)
    text = re.sub(r'bit.ly/\S+', '', text) # remove bitly links
    text = text.strip('[link]') # remove [links]
    return text

df['Adjusted Tweet'] = df['Adjusted Tweet'].apply(remove_links)

## Sentiment Analysis:

Insert Polarity Score Column

df.insert(loc=7,
          column='Polarity Score',
          value=df['Adjusted Tweet'])

Sentiment Analysis using NLTK's VADER

### Uncomment to download lexicon for the first time 
#import nltk
#nltk.download('vader_lexicon')

from nltk.sentiment import SentimentIntensityAnalyzer
sia = SentimentIntensityAnalyzer()

def sentiment_analysis(text):  
    text = sia.polarity_scores(text)
    return text

df['Polarity Score'] = df['Polarity Score'].apply(sentiment_analysis)
    

Creating columns for:
- Negative Score
- Neutral Score
- Positive Score
- Compound Score [-1,1]

df.insert(loc=8,
          column='Negative Score',
          value=df['Polarity Score'])

df.insert(loc=9,
          column='Neutral Score',
          value=df['Polarity Score'])

df.insert(loc=10,
          column='Positive Score',
          value=df['Polarity Score'])

df.insert(loc=11,
          column='Compound Score',
          value=df['Polarity Score'])

df['Negative Score'] = df['Negative Score'].apply(lambda x: x['neg'])
df['Neutral Score'] = df['Neutral Score'].apply(lambda x: x['neu'])
df['Positive Score'] = df['Positive Score'].apply(lambda x: x['pos'])
df['Compound Score'] = df['Compound Score'].apply(lambda x: x['compound'])

df.head()

Sentiment Analysis using TextBlob

df.insert(loc=12,
          column='Polarity Score_textblob',
          value=df['Adjusted Tweet'])

df.insert(loc=13,
          column='Subjectivity Score_textblob',
          value=df['Adjusted Tweet'])

from textblob import TextBlob

#Create a function to get the subjectivity
def getSubjectivity(text):
    return TextBlob(text).sentiment.subjectivity

#Create a function to get the polarity
def getPolarity(text):
    return TextBlob(text).sentiment.polarity

df['Polarity Score_textblob'] = df['Polarity Score_textblob'].apply(getPolarity)
df['Subjectivity Score_textblob'] = df['Subjectivity Score_textblob'].apply(getSubjectivity)


## Further manipulating the tweet

Steps taken:
- Lowercase
- Punctuation
- Tokenization
- Stopword filtering
- Lemmatisation
- Number removal

Changing all text to lowercase

df['Adjusted Tweet'] = df['Adjusted Tweet'].apply(lambda x: x.lower())

Removing all Punctuation

import string

def punctuation_remove(text):
    text = "".join([char for char in text if char not in string.punctuation])
    return text

df['Adjusted Tweet'] = df['Adjusted Tweet'].apply(punctuation_remove)

Number removal

def remove_numbers(text):
    no_numbers = re.sub(r'\d+', '', text)
    return no_numbers

df['Adjusted Tweet'] = df['Adjusted Tweet'].apply(remove_numbers)

Remove other non-ASCII characters

df['Adjusted Tweet'] = df['Adjusted Tweet'].str.replace(r'[^\x00-\x7F]+', '')

Tokenizing

#nltk.download('punkt')

from nltk import word_tokenize

def tokenize(text):
    text = word_tokenize(text)
    return text

df['Adjusted Tweet'] = df['Adjusted Tweet'].apply(tokenize)

Stopword Filtering

from nltk.corpus import stopwords
stop_words = stopwords.words('english')

def remove_stopwords(text):
    text = [word for word in text if word not in stop_words]
    return text

df['Adjusted Tweet'] = df['Adjusted Tweet'].apply(remove_stopwords)

Also remove words of a single letter ('a' and 'I' are already removed)

def remove_singles(text):
    text = [word for word in text if len(word) != 1]
    return text

df['Adjusted Tweet'] = df['Adjusted Tweet'].apply(remove_singles)

Lemmatisation

from nltk.stem import WordNetLemmatizer
wnl = WordNetLemmatizer()

def lemmatise(text): 
    lemma = [wnl.lemmatize(word) for word in text]
    return lemma

df['Adjusted Tweet'] = df['Adjusted Tweet'].apply(lemmatise)

## Months since creation of account relative to tweet

#Finding date of account creation in months
df.insert(loc=5,
          column='Date of Creation in months',
          value=df['user_created'])

from datetime import *

#returning the months of account creation
def account_creation(text):
    text = datetime.strptime(text, "%Y-%m-%d %H:%M:%S+00:00")
    year = str(text)[0:4]
    month = str(text)[5:7]
    total_months = (int(year)*12)+(int(month))
    return (total_months)

df['Date of Creation in months'] = df['Date of Creation in months'].apply(account_creation)

#Finding date of tweet in months
df.insert(loc=6,
          column='Date of Tweet in Months',
          value=df['date'])

#return the year of tweet
df['Date of Tweet in Months'] = df['Date of Tweet in Months'].apply(account_creation)

#Calculating Months since creation of account relative to tweet
df.insert(loc=7,
          column='Months Since Creation of Account',
          value= (df['Date of Tweet in Months']-df['Date of Creation in months']))

## Time of Day

#create a column for hours:
df.insert(loc=2,
          column='hour of tweet',
          value=df['date'])

from datetime import datetime

#return the hour of the tweet
def hour(text):
    text = datetime.strptime(text, "%Y-%m-%d %H:%M:%S+00:00")
    hour = str(text.time())[0:2]
    return int(hour)

df['hour of tweet'] = df['hour of tweet'].apply(hour)

#insert column for the time of day
df.insert(loc=3,
          column='time of day',
          value=df['hour of tweet'])

#calculating the time of day
def time_of_day(text):  
    if ((text >= 4) and (text < 8 )):
        return 'Early Morning'
    elif ((text >= 8) and (text < 12 )):
        return 'Morning'
    elif ((text >= 12) and (text < 16 )):
        return 'Noon'
    elif ((text >= 16) and (text < 20 )):
        return 'Eve'
    elif ((text >= 20) and (text < 24 )):
        return 'Night'
    elif ((text >= 0) and (text < 4 )):
        return 'Late Night'
    
df['time of day'] = df['time of day'].apply(time_of_day)


#Creating counts using one hot encoding

#Early Morning Count
df.insert(loc=4,
          column='Early Morning Count',
          value=df['time of day'])

def early_morning_count(text):
    if text == 'Early Morning':
        return 1
    else:
        return 0

df['Early Morning Count'] = df['Early Morning Count'].apply(early_morning_count)

#Morning Count
df.insert(loc=5,
          column='Morning Count',
          value=df['time of day'])

def morning_count(text):
    if text == 'Morning':
        return 1
    else:
        return 0

df['Morning Count'] = df['Morning Count'].apply(morning_count)

#Noon count
df.insert(loc=6,
          column='Noon Count',
          value=df['time of day'])

def noon_count(text):
    if text == 'Noon':
        return 1
    else:
        return 0

df['Noon Count'] = df['Noon Count'].apply(noon_count)

#Eve count
df.insert(loc=7,
          column='Eve Count',
          value=df['time of day'])

def eve_count(text):
    if text == 'Eve':
        return 1
    else:
        return 0

df['Eve Count'] = df['Eve Count'].apply(eve_count)

#Night count
df.insert(loc=8,
          column='Night Count',
          value=df['time of day'])

def night_count(text):
    if text == 'Night':
        return 1
    else:
        return 0

df['Night Count'] = df['Night Count'].apply(night_count)

#Late Night count
df.insert(loc=9,
          column='Late Night Count',
          value=df['time of day'])

def late_night_count(text):
    if text == 'Late Night':
        return 1
    else:
        return 0

df['Late Night Count'] = df['Late Night Count'].apply(late_night_count)

## Video, GIF and Photo Count

Photo Count

#Creating a photo count column...
df.insert(loc=23,
          column='Photo Count',
          value=df['media'])

#Counting number of Photos in media column
#No need to tokenize
def photo_count(text):
    text = str(text)
    text = text.count('Photo')
    return text

df['Photo Count'] = df['Photo Count'].apply(photo_count)


Video Count

#Creating a video count column...
df.insert(loc=24,
          column='Video Count',
          value=df['media'])

#We need to tokenize the media column so that we can count how many videos there are...
from nltk import word_tokenize

def tokenize(text):
    text = str(text)
    text = word_tokenize(text)
    return text

df['Video Count'] = df['Video Count'].apply(tokenize)

#Counting number of Videos in media column
def video_count(text):
    text = text.count('Video')
    return text

df['Video Count'] = df['Video Count'].apply(video_count)

Gif Count

#Creating a GIF count column...
df.insert(loc=25,
          column='GIF Count',
          value=df['media'])

#We need to tokenize the media column so that we can count how many GIFs there are...
from nltk import word_tokenize

def tokenize(text):
    text = str(text)
    text = word_tokenize(text)
    return text

df['GIF Count'] = df['GIF Count'].apply(tokenize)

#Counting number of GIFs in media column
def gif_count(text):
    text = text.count('Gif')
    return text

df['GIF Count'] = df['GIF Count'].apply(gif_count)

## Topic Modelling

from sklearn.feature_extraction.text import CountVectorizer

def dummy(x):
    return x

vectorizer = CountVectorizer(max_df = 0.9, min_df = 25, lowercase = False, tokenizer = dummy)
tf = vectorizer.fit_transform(df['Adjusted Tweet']).toarray()
tf_features_names = vectorizer.get_feature_names()

from sklearn.decomposition import LatentDirichletAllocation

number_of_topics = 3

model = LatentDirichletAllocation(n_components=number_of_topics)

model.fit(tf)

def display_topics(model, feature_names, no_top_words):
    topic_dict = {}
    for topic_idx, topic in enumerate(model.components_):
        topic_dict["Topic %d words" % (topic_idx)]= ['{}'.format(feature_names[i])
                        for i in topic.argsort()[:-no_top_words - 1:-1]]
        topic_dict["Topic %d weights" % (topic_idx)]= ['{:.1f}'.format(topic[i])
                        for i in topic.argsort()[:-no_top_words - 1:-1]]
    return pd.DataFrame(topic_dict)

no_top_words = 10
display_topics(model, tf_feature_names, no_top_words)

from sklearn.decomposition import NMF

model_2 = NMF(n_components=4, random_state=0, alpha=.1, l1_ratio=.5)

model_2.fit(tf)

display_topics(model_2, tf_feature_names, no_top_words)

Seems like these clustering algorithms do not pick up any interesting clusters

## Creating our final Dataframe

Drop unnecessary features

final_df = df.drop(['media','Polarity Score','user_created','Date of Creation in months','Date of Tweet in Months',
         'hashtags', 'lang', 'mentionedUsers'], axis=1)

final_df.head()

final_df.info()

Final data cleaning:

final_df.to_csv('../Data/Processed Dataset.csv', index = False)
