==> Project Code/3. Word Vectorisation.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv("../Data/Processed Dataset.csv")

df.columns

## TF-IDF Vectorisation

import ast
df["Adjusted Tweet"] = df["Adjusted Tweet"].apply(ast.literal_eval)

from sklearn.feature_extraction.text import TfidfVectorizer
def dummy(x):
    return x

vectorizer = TfidfVectorizer(max_df = 0.9, min_df = 50, lowercase = False, tokenizer = dummy)
tf = vectorizer.fit_transform(df['Adjusted Tweet']).toarray()
tf_feature_names = vectorizer.get_feature_names()

df_tf_idf = pd.DataFrame(tf, columns = tf_feature_names)
df_tf_idf["id"] = df["id"]

df_tf_idf.head()

df_tfidf_merged = pd.merge(df, df_tf_idf, on = "id")

df_tfidf_merged

df_tfidf_merged.to_csv('../Data/Processed Dataset TF-IDF.csv', index = False)

df_tfidf_merged.info()

## Word2vec

from gensim.models import Word2Vec

corpus = df["Adjusted Tweet"].to_numpy()
corpus

model = Word2Vec(min_count = 50)
model.build_vocab(corpus)

# idea: replace words not in the vocabulary with an unknown word token and train with it
# due to some tweets only containing 'unique' words, so do not have a document embedding
def unknowns(doc):
    unknown_token = "xxunkxx"
    l = [word if word in model.wv.key_to_index else unknown_token for word in doc]
    if not l:
        l = [unknown_token]
    return l

df["Adjusted Unknowns"] = df["Adjusted Tweet"].apply(unknowns)
corpus_unk = df["Adjusted Unknowns"].to_numpy()

model = Word2Vec(min_count = 50)
model.build_vocab(corpus_unk) #use new corpus to train
model.train(corpus_unk, total_examples = model.corpus_count, epochs = 10)

embeddings = model.wv
len(embeddings.key_to_index)

embeddings.similar_by_word("xxunkxx")

Create vector representations

def doc_vector(doc):
    l = pd.Series([word for word in doc if word in embeddings.index_to_key], dtype = 'object')
    l = l.apply(embeddings.get_vector)
    return np.mean(l, axis = 0)

doc_embedding_matrix = np.array(df["Adjusted Unknowns"].apply(doc_vector).tolist())
np.shape(doc_embedding_matrix)

column_names = []
for i in range(100):
    column_names.append(f"doc_embed_{i}")

df_doc_embeddings = pd.DataFrame(doc_embedding_matrix, columns = column_names)

df_doc_embeddings["id"] = df["id"]
df_embeddings_merged = pd.merge(df, df_doc_embeddings, on = 'id')

df_embeddings_merged.info()

df_embeddings_merged.to_csv('../Data/Processed Dataset Doc Embeddings.csv', index = False)

def cosine_similarity(x, y):
    
    # Ensure length of x and y are the same
    if len(x) != len(y) :
        return None
    
    # Compute the dot product between x and y
    dot_product = np.dot(x, y)
    
    # Compute the L2 norms (magnitudes) of x and y
    magnitude_x = np.sqrt(np.sum(x**2)) 
    magnitude_y = np.sqrt(np.sum(y**2))
    
    # Compute the cosine similarity
    cosine_similarity = dot_product / (magnitude_x * magnitude_y)
    
    return cosine_similarity

cosine_similarity(doc_embedding_matrix[0], doc_embedding_matrix[4])
