==> Project Code/4. Dataset Split.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split

path = '../Data/'
datasets = ['Processed Dataset', 'Processed Dataset Doc Embeddings', 'Processed Dataset TF-IDF']

for name in datasets:
    df = pd.read_csv(path + name + '.csv')
    df_train, df_test = train_test_split(df, test_size = 0.2, random_state = 0)
    print(path + name)
    df_train.to_csv(path + name + ' Train', index = False)
    df_test.to_csv(path + name + ' Test', index = False)
